There is no inotify on OpenBSD
The Dart documentation on DirectoryWatcher states:
  /// If a native directory watcher is available for this platform, this will
  /// use it. Otherwise, it will fall back to a [PollingDirectoryWatcher].

This leads me to believe that we are not forced to implement
the functionality & kqueue would require us to open a file handle
for all observed files inside a directory.

The check is done by if (FileSystemEntity.isWatchSupported) {
and then instantinates a dart 4 platform class ie. LinuxDirectoryWatcher.

We will probably need to later on implement the isWatchSupported, which
translates to isSupported call on the class.

For now skip the file in the build.